+++ pkg/okteto/kubetoken/cache_test.go
@@ -1,0 +1,17 @@
+package kubetoken
+
+import "testing"
+
+func TestFileCache(t *testing.T) {
+
+	// Test file does not exist
+
+	// Test file has corrupted data
+
+	// Test error while writing file
+
+	// Test error while stat file
+
+	// Test token expiration
+
+}
