+++ pkg/okteto/kubetoken/kubetoken_client.go
@@ -15,12 +15,9 @@ package kubetoken
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
-	"os"
-	"time"
 
 	authenticationv1 "k8s.io/api/authentication/v1"
 
@@ -34,115 +31,6 @@ type storeRegister struct {
 	ContextName string                        `json:"context"`
 	Namespace   string                        `json:"namespace"`
 	Token       authenticationv1.TokenRequest `json:"token"`
-}
-
-type FileByteStore struct {
-	FileName string
-}
-
-func (s *FileByteStore) Get() ([]byte, error) {
-	if _, err := os.Stat(s.FileName); err != nil {
-		if !errors.Is(err, os.ErrNotExist) {
-			return nil, fmt.Errorf("error checking if file exists: %w", err)
-		}
-
-		if err := os.WriteFile(s.FileName, []byte("[]"), 0600); err != nil {
-			return nil, fmt.Errorf("error creating file: %w", err)
-		}
-	}
-
-	contents, err := os.ReadFile(s.FileName)
-	if err != nil {
-		return nil, fmt.Errorf("error reading file: %w", err)
-	}
-
-	return contents, nil
-}
-
-func (s *FileByteStore) Set(value []byte) error {
-	return os.WriteFile(s.FileName, value, 0600)
-}
-
-type stringStore interface {
-	Get() ([]byte, error)
-	Set([]byte) error
-}
-
-type Cache struct {
-	StringStore stringStore
-}
-
-func (c *Cache) read() ([]storeRegister, error) {
-	contents, err := c.StringStore.Get()
-	if err != nil {
-		return nil, err
-	}
-
-	var store []storeRegister
-
-	if err := json.Unmarshal(contents, &store); err != nil {
-		return nil, fmt.Errorf("error decoding") // TODO: we should probably delete the file contents
-	}
-
-	return store, nil
-}
-
-func (c *Cache) Get(contextName, namespace string) (string, error) {
-	store, err := c.read()
-	if err != nil {
-		return "", err
-	}
-
-	for _, register := range store {
-		if register.ContextName == contextName && register.Namespace == namespace {
-			now := time.Now() // TODO: inject this
-			if register.Token.Status.ExpirationTimestamp.Time.After(now) {
-				tokenString, _ := json.MarshalIndent(register.Token, "", "\t")
-
-				return string(tokenString), nil
-			} else {
-				// TODO: we could invalidate this cache here
-				return "", nil
-			}
-		}
-	}
-
-	return "", nil
-}
-
-func (c *Cache) setWithErr(contextName, namespace string, token authenticationv1.TokenRequest) error {
-	store, err := c.read()
-	if err != nil {
-		return err
-	}
-
-	existed := false
-	for i, r := range store {
-		if r.ContextName == contextName && r.Namespace == namespace {
-			store[i].Token = token
-			existed = true
-		}
-	}
-	if !existed {
-		store = append(store, storeRegister{
-			ContextName: contextName,
-			Namespace:   namespace,
-			Token:       token,
-		})
-	}
-
-	newStore, err := json.MarshalIndent(store, "", "\t")
-	if err != nil {
-		return err
-	}
-
-	return c.StringStore.Set(newStore)
-}
-
-func (c *Cache) Set(contextName, namespace string, token authenticationv1.TokenRequest) {
-	if err := c.setWithErr(contextName, namespace, token); err != nil {
-		// TODO: log this
-	}
 }
 
 type cacheSetter interface {
